point mixer at gemini jobcoin

+++ api/src/main/scala/lavador/mixer/Model.scala
@@ -10,7 +10,7 @@ case class MaxMixerFee(value: Double) extends AnyVal
 case class MaxTimeToExecute(value: FiniteDuration) extends AnyVal
 
 case class MixerReceipt(
-  fee: Int,
+  fee: BigDecimal,
   scheduledDepositTime: ZonedDateTime,
   transactions: List[Transaction]
 )
